requirements.txt
appdirs==1.4.4
asn1crypto==0.24.0
configparser==4.0.2
contextlib2==0.6.0.post1
cryptography==2.1.4
distlib==0.3.0
enum34==1.1.6
filelock==3.0.12
idna==2.6
importlib-metadata==1.6.0
importlib-resources==1.5.0
ipaddress==1.0.17
keyring==10.6.0
keyrings.alt==3.0
pathlib2==2.3.5
pycrypto==2.6.1
pygobject==3.26.1
pytz==2020.1
pyxdg==0.25
scandir==1.10.0
SecretStorage==2.3.1
singledispatch==3.4.0.3
six==1.11.0
typing==3.7.4.1
virtualenv==20.0.21
zipp==1.2.0
Django==2.1.3
django-crispy-forms==1.7.2
